+++ thewalrus/decompositions.py
@@ -152,7 +152,7 @@ def blochmessiah(S):
     return O, D, Q
 
 
-def takagi(A, svd_order=True):
+def takagi(A, svd_order=True, tol=1e-16):
     r"""Autonne-Takagi decomposition of a complex symmetric (not Hermitian!) matrix.
     Note that the input matrix is internally symmetrized by taking its upper triangular part.
     If the input matrix is indeed symmetric this leaves it unchanged.
@@ -162,6 +162,7 @@ def takagi(A, svd_order=True):
     Args:
         A (array): square, symmetric matrix
         svd_order (boolean): whether to return result by ordering the singular values of ``A`` in descending (``True``) or ascending (``False``) order.
+        tol (float): tolerance parameter to kill the small values in the maatix A. This was installed because takagi decomposition did not work at some matrix A.
 
     Returns:
         tuple[array, array]: (r, U), where r are the singular values,
@@ -174,7 +175,7 @@ def takagi(A, svd_order=True):
     # Here we build a Symmetric matrix from the top right triangular part
     A = np.triu(A) + np.triu(A, k=1).T
 
-    A = np.real_if_close(A)
+    A = np.where(np.abs(A.real) <= tol, 0, A.real) + np.where(np.abs(A.imag) <= tol, 0, A.imag)
 
     if np.allclose(A, 0):
         return np.zeros(n), np.eye(n)